Implemented rk4ip method.
Modified split step method so it takes the same args as RK4IP, as well
as being able to take the nonlocal operator as an array in Fourier space
instead of only an OperatorSum instance.

Modified plot_example.py to transpose the image so that x and y directions
are correct.